Move object stores into dedicated module
